requirements.txt
docker==7.1.0
pyyaml==6.0.2
click==8.1.7
rich==13.9.4
psutil==6.1.0
requests==2.32.3
mcp==1.1.0
